+++ packages/sdk/src/types.ts
@@ -792,7 +792,7 @@ export interface ChainArData {
 export interface Policy {
     validTo: string; // "YYYYMM"
     createdAt: string; // "YYYYMM"
-    revealedAttributes: Record<AttributeKey, string>;
+    revealedAttributes: Partial<Record<AttributeKey, string>>;
 }
 
 interface SharedCredentialDeploymentValues {
@@ -818,7 +818,7 @@ export interface CredentialDeploymentCommitments {
     cmmPrf: string;
     cmmCredCounter: string;
     cmmIdCredSecSharingCoeff: string[];
-    cmmAttributes: Record<AttributeKey, string>;
+    cmmAttributes: Partial<Record<AttributeKey, string>>;
     cmmMaxAccounts: string;
 }
 
@@ -1600,7 +1600,7 @@ export interface AttributeList {
     validTo: string; // "YYYYMM"
     createdAt: string; // "YYYYMM"
     maxAccounts: number;
-    chosenAttributes: Record<AttributeKey, string>;
+    chosenAttributes: Partial<Record<AttributeKey, string>>;
 }
 
 export type IdentityObjectV1 = {